HZN-322 return node graph definitions and resources in one call
