+++ opennms-webapp-rest/src/main/java/org/opennms/web/rest/GraphNameResolver.java
@@ -1,0 +1,5 @@
+package org.opennms.web.rest;
+
+public interface GraphNameResolver {
+    GraphNameCollection getGraphNames(final String resourceId);
+}
