+++ opennms-webapp-rest/src/main/java/org/opennms/web/rest/GraphRestService.java
@@ -28,22 +28,27 @@
 
 package org.opennms.web.rest;
 
-import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 
+import javax.ws.rs.DefaultValue;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
+import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response.Status;
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 
-import org.opennms.core.config.api.JaxbListWrapper;
 import org.opennms.netmgt.dao.api.GraphDao;
+import org.opennms.netmgt.dao.api.NodeDao;
 import org.opennms.netmgt.dao.api.ResourceDao;
+import org.opennms.netmgt.model.OnmsNode;
 import org.opennms.netmgt.model.OnmsResource;
 import org.opennms.netmgt.model.PrefabGraph;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -53,6 +58,7 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.google.common.collect.Lists;
+import com.google.common.collect.Maps;
 
 /**
  * Read-only API for retrieving graph definitions and determining
@@ -64,6 +70,9 @@ import com.google.common.collect.Lists;
 @Scope("prototype")
 @Path("graphs")
 public class GraphRestService extends OnmsRestService {
+
+    @Autowired
+    private NodeDao m_nodeDao;
 
     @Autowired
     private GraphDao m_graphDao;
@@ -115,22 +124,78 @@ public class GraphRestService extends OnmsRestService {
         return new GraphNameCollection(graphNames);
     }
 
-    @XmlRootElement(name = "names")
-    public static final class GraphNameCollection extends JaxbListWrapper<String> {
+    @GET
+    @Path("fornode/{nodeCriteria}")
+    @Produces({MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON, MediaType.APPLICATION_ATOM_XML})
+    @Transactional(readOnly=true)
+    public GraphResourceDTO getGraphResourcesForNode(@PathParam("nodeCriteria") final String nodeCriteria, @DefaultValue("-1") @QueryParam("depth") final int depth) {
+        OnmsNode node = m_nodeDao.get(nodeCriteria);
+        if (node == null) {
+            throw getException(Status.NOT_FOUND, "No node found with criteria '{}'.", nodeCriteria);
+        }
 
-        private static final long serialVersionUID = 1L;
+        OnmsResource resource = m_resourceDao.getResourceForNode(node);
+        if (resource == null) {
+            throw getException(Status.NOT_FOUND, "No resource found for node with id {}.", "" + node.getId());
+        }
 
-        public GraphNameCollection() {
+        final ResourceVisitor visitor = new ResourceVisitor(this);
+        final ResourceDTO resourceDTO = new ResourceDTO(resource, depth);
+        visitor.visit(resourceDTO);
+
+        return new GraphResourceDTO(resourceDTO, visitor.getGraphs());
+    }
+
+    @XmlRootElement(name = "graph-resource")
+    @XmlAccessorType(XmlAccessType.NONE)
+    public static final class GraphResourceDTO {
+        @XmlElement(name="resource")
+        private ResourceDTO m_resource;
+
+        @XmlElement(name="prefab-graphs")
+        private PrefabGraphCollection m_prefabGraphs;
+
+        @SuppressWarnings("unused")
+        private GraphResourceDTO() {
             super();
         }
 
-        public GraphNameCollection(Collection<? extends String> names) {
-            super(names);
+        public GraphResourceDTO(final ResourceDTO resource, final PrefabGraphCollection graphs) {
+            m_resource = resource;
+            m_prefabGraphs = graphs;
+        }
+    }
+
+    public static final class ResourceVisitor {
+        private final Map<String,PrefabGraph> m_graphs = Maps.newLinkedHashMap();
+        private final GraphRestService m_service;
+
+        public ResourceVisitor(GraphRestService service) {
+            m_service = service;
         }
 
-        @XmlElement(name = "name")
-        public List<String> getObjects() {
-            return super.getObjects();
+        public void visit(final ResourceDTO resource) {
+            // first, decorate the DTO with the list of graph names
+            final GraphNameCollection graphNames = m_service.getGraphNamesForResource(resource.getId());
+            resource.setGraphNames(graphNames.getObjects());
+
+            // then, get the prefab graphs for these graph names if we don't have them already
+            for (final String graphName : graphNames) {
+                if (!m_graphs.containsKey(graphName)) {
+                    m_graphs.put(graphName, m_service.getGraphByName(graphName));
+                }
+            }
+
+            // finally, recurse if necessary
+            if (resource.getChildren() != null) {
+                for (final ResourceDTO r : resource.getChildren()) {
+                    this.visit(r);
+                }
+            }
+        }
+
+        public PrefabGraphCollection getGraphs() {
+            return new PrefabGraphCollection(m_graphs.values());
         }
     }
 }
